th-engine: fused add+rmsnorm, bf16 verify readback
One add_rmsnorm dispatch per layer boundary emits both the residual
and normed streams (was add + separate norm — ~128 fewer dispatches
per forward). forward_multi returns raw bf16 logits; the accept pass
converts per-row on CPU, halving the [seq,vocab] readback and skipping
the GPU convert.

+++ engine/src/main.rs
@@ -299,7 +299,8 @@ async fn main() -> Result<()> {
                     let t = std::time::Instant::now();
                     let lg =
                         loaded.backend.forward_multi(&seq, pos, &dev)?;
-                    let _ = lg.flatten_all()?.to_vec1::<f32>()?;
+                    let _ =
+                        lg.flatten_all()?.to_vec1::<half::bf16>()?;
                     eprintln!(
                         "fwd{m}  {:.1}ms",
                         t.elapsed().as_secs_f64() * 1e3
